+++ src/unistd.rs
@@ -872,6 +872,36 @@ pub fn lseek64(fd: RawFd, offset: libc::off64_t, whence: Whence) -> Result<libc:
     Errno::result(res).map(|r| r as libc::off64_t)
 }
 
+/// Call the link function to create a link to a file
+/// ([posix specification](http://pubs.opengroup.org/onlinepubs/9699919799/functions/link.html)).
+pub fn link<P1: ?Sized + NixPath, P2: ?Sized + NixPath>(oldpath: &P1, newpath: &P2) -> Result<()> {
+    let res = try!(try!(oldpath.with_nix_path(|old|
+        newpath.with_nix_path(|new|
+            unsafe {
+                libc::link(old.as_ptr() as *const c_char, new.as_ptr() as *const c_char)
+            }
+        )
+    )));
+
+    Errno::result(res).map(drop)
+}
+
+/// Call the link function to create a link to a file
+/// ([posix specification](http://pubs.opengroup.org/onlinepubs/9699919799/functions/linkat.html)).
+pub fn linkat<P1: ?Sized + NixPath, P2: ?Sized + NixPath>(olddirfd: RawFd, oldpath: &P1,
+                                                          newdirfd: RawFd, newpath: &P2, flags: AtFlags) -> Result<()> {
+    let res = try!(try!(oldpath.with_nix_path(|old|
+        newpath.with_nix_path(|new|
+            unsafe {
+                libc::linkat(olddirfd, old.as_ptr() as *const c_char,
+                             newdirfd, new.as_ptr() as *const c_char, flags.bits())
+            }
+        )
+    )));
+
+    Errno::result(res).map(drop)
+}
+
 /// Create an interprocess channel.
 ///
 /// See also [pipe(2)](http://pubs.opengroup.org/onlinepubs/9699919799/functions/pipe.html)