+++ CHANGELOG.md
@@ -61,6 +61,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   ([#753](https://github.com/nix-rust/nix/pull/753))
 - Added `nix::unistd::mkdirat`
   ([#754](https://github.com/nix-rust/nix/pull/754))
+- Added `nix::unistd::{link, linkat}`
+  ([#755](https://github.com/nix-rust/nix/pull/755))
 
 ### Changed
 - Use native `pipe2` on all BSD targets.  Users should notice no difference.